Initial commit after iteration with GPT-4
Interestingly, GPT-4 output code that contains a subtle JS bug, due to
the way closures work in JS. Notice that 'li' element is defined as a
`var`, which has no block scope and effectively acts like a global
variable in this case (see https://javascript.info/var). So when we
click the button, it alway adds the spinner and error message on the
last list item!